- Update parser_task to have last ditch exception handling - Move parsing logging to util file
